pyproject.toml
[project]
name = "python1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.0.15",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "numpy>=2.2.6",
    "openai>=1.0.0",
    "openi>=2.0.4",
    "pandas>=2.3.1",
    "pymupdf>=1.26.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "tiktoken>=0.9.0",
]
